+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+- On X11 and Wayland, you can now use shared contexts, however, one limitation 
+of the Wayland backend is that all shared contexts must use the same events
+pool as each other.
 - ***Breaking*** The entire api for Headless contexts have been removed. Please instead use `Context::new()` when trying to make a context without a visible window. Also removed `headless` feature.
 - ***Breaking*** Structs implementing the `GlContext` trait must now be sized.
 - ***Breaking*** Added new `CreationErrorPair` enum variant to enum `CreationError`.